refactor(compiler-cli): refactor import adding logic into helper
This logic will be common to transforms that add imports. Using it as a helper helps reduce duplication
